python-scripts/requirements.txt
psycopg2-binary==2.9.9
Faker==24.4.0